feat: feature 레이어로 query 이동
