+++ src/features/index.ts
@@ -1,2 +1,6 @@
 export * as AuthFeature from './auth';
 export * as BoardFeature from './board';
+export * as FollowFeature from './follow';
+export * as MarketFeature from './market';
+export * as MemberFeature from './member';
+export * as TickerFeature from './ticker';
