+++ src/features/market/model/query.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { options } from './query-options';
+
+import type { MarketType, MarketCandleRange } from '@shared';
+
+export const useSelectMarketsQuery = (marketType: MarketType) => useQuery(options.selectFilteredMarkets(marketType));
+
+export const useMarketCandlesMinuteQuery = (market: string, range: MarketCandleRange) =>
+  useQuery(options.getMarketCandlesWithUnit({ market, range }));
