+++ src/features/ticker/model/query.ts
@@ -1,0 +1,5 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { options } from './query-options';
+
+export const useMarketTickerQuery = (market: string) => useQuery(options.getMarketTicker(market));
