--- src/queries/board.ts
@@ -1,15 +1,0 @@
-import { queryOptions } from "@tanstack/react-query";
-
-const options = {
-    rootKey: 'board',
-    getIdeas: () =>
-      queryOptions({
-        queryKey: [options.rootKey, 'ideas'],
-        queryFn: () => MarketService.getMarkets(),
-      }),
-    getFeeds: () =>
-      queryOptions({
-        queryKey: [options.rootKey, 'feeds'],
-        queryFn: () => MarketService.getMarkets(),
-      }),
-  };
